+++ src/layouts/modal/ModalLayout.jsx
@@ -1,9 +1,11 @@
 import { createPortal } from "react-dom";
-import * as S from "./modal.style.js";
+import * as S from "./modalLayout.style.js";
 
 export default function Modal({ closeModal, children }) {
   return createPortal(
-    <S.ModalWrapper onClick={closeModal}>{children}</S.ModalWrapper>,
+    <S.ModalLayoutWrapper onClick={closeModal}>
+      {children}
+    </S.ModalLayoutWrapper>,
     document.getElementById("modal-root")
   );
 }
